+++ examples/1.1-check-solarium-and-ping.php
@@ -1,6 +1,6 @@
 <?php
 
-require(__DIR__.'/init.php');
+require_once(__DIR__.'/init.php');
 htmlHeader();
 
 // check solarium version available