+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 ### Added
+- \Solarium\Support\Utility::getXmlEncoding()
+
+### Fixed
+- MoreLikeThis fails on Solr Cloud
+- MinimumScoreFIlter Plugin might fail on Solr 7 in cloud mode
 
 ### Changed